backend/requirements.txt
fastapi
uvicorn
python-multipart
yt-dlp
openai-whisper
moviepy
google-generativeai
groq
opencv-python
Pillow
streamlit
requests
ffmpeg-python
numpy<2
torch